+++ src/main/java/uk/co/real_logic/agrona/concurrent/AtomicBuffer.java
@@ -35,7 +35,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes from which to get.
      * @return the value for at a given index
      */
-    long getLongVolatile(int index);
+    long getLongVolatile(long index);
 
     /**
      * Put a value to a given index with volatile semantics.
@@ -43,7 +43,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putLongVolatile(int index, long value);
+    void putLongVolatile(long index, long value);
 
     /**
      * Put a value to a given index with ordered store semantics.
@@ -51,7 +51,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putLongOrdered(int index, long value);
+    void putLongOrdered(long index, long value);
 
     /**
      * Add a value to a given index with ordered store semantics. Use a negative increment to decrement.
@@ -59,7 +59,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param increment by which the value at the index will be adjusted.
      * @return the previous value at the index
      */
-    long addLongOrdered(int index, long increment);
+    long addLongOrdered(long index, long increment);
 
     /**
      * Atomic compare and set of a long given an expected value.
@@ -69,7 +69,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param updateValue   to be exchanged
      * @return set successful or not
      */
-    boolean compareAndSetLong(int index, long expectedValue, long updateValue);
+    boolean compareAndSetLong(long index, long expectedValue, long updateValue);
 
     /**
      * Atomically exchange a value at a location returning the previous contents.
@@ -78,7 +78,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param value for at a given index
      * @return previous value at the index
      */
-    long getAndSetLong(int index, long value);
+    long getAndSetLong(long index, long value);
 
     /**
      * Atomically add a delta to a value at a location returning the previous contents.
@@ -88,7 +88,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param delta to be added to the value at the index
      * @return previous value
      */
-    long getAndAddLong(int index, long delta);
+    long getAndAddLong(long index, long delta);
 
     /**
      * Get the value at a given index with volatile semantics.
@@ -96,7 +96,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes from which to get.
      * @return the value for at a given index
      */
-    int getIntVolatile(int index);
+    int getIntVolatile(long index);
 
     /**
      * Put a value to a given index with volatile semantics.
@@ -104,7 +104,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putIntVolatile(int index, int value);
+    void putIntVolatile(long index, int value);
 
     /**
      * Put a value to a given index with ordered semantics.
@@ -112,7 +112,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putIntOrdered(int index, int value);
+    void putIntOrdered(long index, int value);
 
     /**
      * Add a value to a given index with ordered store semantics. Use a negative increment to decrement.
@@ -120,7 +120,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param increment by which the value at the index will be adjusted.
      * @return the previous value at the index
      */
-    int addIntOrdered(int index, int increment);
+    int addIntOrdered(long index, int increment);
 
     /**
      * Atomic compare and set of a int given an expected value.
@@ -130,7 +130,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param updateValue   to be exchanged
      * @return successful or not
      */
-    boolean compareAndSetInt(int index, int expectedValue, int updateValue);
+    boolean compareAndSetInt(long index, int expectedValue, int updateValue);
 
     /**
      * Atomically exchange a value at a location returning the previous contents.
@@ -139,7 +139,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param value for at a given index
      * @return previous value
      */
-    int getAndSetInt(int index, int value);
+    int getAndSetInt(long index, int value);
 
     /**
      * Atomically add a delta to a value at a location returning the previous contents.
@@ -149,7 +149,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param delta to be added to the value at the index
      * @return previous value
      */
-    int getAndAddInt(int index, int delta);
+    int getAndAddInt(long index, int delta);
 
     /**
      * Get the value at a given index with volatile semantics.
@@ -157,7 +157,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes from which to get.
      * @return the value for at a given index
      */
-    short getShortVolatile(int index);
+    short getShortVolatile(long index);
 
     /**
      * Put a value to a given index with volatile semantics.
@@ -165,7 +165,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putShortVolatile(int index, short value);
+    void putShortVolatile(long index, short value);
 
     /**
      * Get the value at a given index with volatile semantics.
@@ -173,7 +173,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes from which to get.
      * @return the value for at a given index
      */
-    char getCharVolatile(int index);
+    char getCharVolatile(long index);
 
     /**
      * Put a value to a given index with volatile semantics.
@@ -181,7 +181,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putCharVolatile(int index, char value);
+    void putCharVolatile(long index, char value);
 
     /**
      * Get the value at a given index with volatile semantics.
@@ -189,7 +189,7 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes from which to get.
      * @return the value for at a given index
      */
-    byte getByteVolatile(int index);
+    byte getByteVolatile(long index);
 
     /**
      * Put a value to a given index with volatile semantics.
@@ -197,5 +197,5 @@ public interface AtomicBuffer extends MutableDirectBuffer
      * @param index in bytes for where to put.
      * @param value for at a given index
      */
-    void putByteVolatile(int index, byte value);
+    void putByteVolatile(long index, byte value);
 }
